requirements.txt
ai-api-client-sdk==2.2.1
ai-core-sdk==2.4.5
generative-ai-hub-sdk==3.1.1
cfenv==0.5.3
sap-xssec
Flask
gunicorn
langchain==0.2.16
langchain-community==0.2.10
langchain-core==0.2.41
langchain-google-genai==1.0.10
langchain_google_vertexai==1.0.10
langchain-openai==0.1.20
langchain-text-splitters==0.2.0
langchain_aws==0.1.17
google-generativeai==0.7.2
